+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 * Joker, JokerCast, JokerFrame, JokerListener, JokerPortal, JokerStage, JokerTroupe (State
   Management)
 * CircusRing (Dependency Injection)
+
+## 1.0.1
+
+* feat: add fenix support to CircusRing
+* Add examples/tests